Main code of our project
This is the code where we connect the previous database file via using the connector driver of python. Then we bring the question's from the database to to code in list and dictionary. Then we perform to copy main question and shuffle the questions with random function. then we copied as like pervious one by one set how much number of set we want.